requirements.txt
asgiref==3.5.0
Django==4.0.2
django-fernet-fields==0.6
gunicorn==20.1.0
psycopg2-binary==2.8.6
pytz==2021.1
sqlparse==0.4.1
selenium==4.1.0
Scrapy==2.5.0
Twisted==22.1.0
whitenoise==5.2.0
whitenoise==5.2.0
django-fernet-fields==0.6
django-nepali-datetime-field==0.4
mechanize==0.4.7
beautifulsoup4==4.10.0
urllib3==1.26.6
httplib2==0.20.4
pyBSDate==0.3.0
Pillow==9.0.1
djangorestframework==3.13.1
requests==2.28.1